Initial commit
Created starting point

Finished

Made click events all have parens

Removed absolute path to css and js

Update README.md

Fixed bug when changing repos

There was a bug found that would create a 404 when switching repos if the path was not the root.  It would try to request a path in the new repo that possibly wouldn't exist.  Simply just need to make the path the root path before requesting the new files.