Change PageNavBtns component to use prop functions
Also updates Vue to latest version and simplifies the Menu
component.

Pass down parent functions as props instead of hardcoding
this.$parent.function in child component.

In case the parent function is changed or otherwise defined
differently one could just modify which function is passed
down. This eliminates the need to update the child component
to accomodate for changes or differences in parent components
in the future.