Implement Vuex store to replace event system
Use Vuex store to maintain a single source of application state
(e.g. error messages and username/repo values). Doing this loosens
up the coupling between the main app component and the children,
as we no longer need to pass events between the components.

Events will also be deprecated in Vue >= 2.0 (which is part of the reason
for these changes).